+++ src/icons/ChartDonut2Icon.tsx
@@ -1,0 +1,20 @@
+import { wrapIcon } from './wrap-icon';
+
+export const ChartDonut2Icon = wrapIcon(
+  'ChartDonut2Icon',
+  <svg
+    viewBox="0 0 24 24"
+    xmlns="http://www.w3.org/2000/svg"
+    width="24"
+    height="24"
+    fill="none"
+  >
+    <path
+      stroke="#000"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth="1.5"
+      d="M12 3v5m0-5a9 9 0 0 1 9 9m-9-9a9 9 0 1 0 9 9m-9-4a4 4 0 0 1 4 4m-4-4a4 4 0 1 0 4 4m0 0h5"
+    />
+  </svg>,
+);